% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_tools.R
\name{changepos}
\alias{changepos}
\title{Get list of indices corresponding to site the mutation pattern changes}
\usage{
changepos(mutpats)
}
\description{
Get list of indices corresponding to site the mutation pattern changes
}
