% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_residuals.R
\name{clusterresids}
\alias{clusterresids}
\title{Look for motifs in residuals}
\usage{
clusterresids(resids, npats = 300, nclusts = 12, top = TRUE,
  return.list = FALSE, ...)
}
\arguments{
\item{resids}{A table of residuals containing 'inpat', 'outpat' and a 'z' score.}

\item{npats}{The number of top patterns to look for motifs in.}

\item{nclusts}{The number of clusters to report (using hclust).}

\item{top}{Whether to do those with large positive z scores? Otherwise, negative.}

\item{return.list}{Whether to output the full list of patterns in each motif.}

\item{...}{Additional arguments passed to print.motif.}
}
\description{
For the top and bottom sets of `npats` residuals, does the following:
 - pastes 'long' and 'short' sequences together
 - finds the matrix of hamming distances between these
 - uses hclust to cluster the results
 - uses hclust to cluster the results
}
