% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons.R
\docType{data}
\name{codons}
\alias{codons}
\title{Codons}
\format{A data frame with columns 'codon', 'aacode', 'aa', and 'regexp', like
   TTT  F  Phe  TT[TC]}
\usage{
codons
}
\description{
Codons
}
\keyword{datasets}
