% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_estimates.R
\name{countmuts}
\alias{countmuts}
\title{Count Numbers of Mutations}
\usage{
countmuts(counts, mutpats, leftwin = leftwin(counts), ...)
}
\description{
Given a contingency table `counts` of Tmers observed in a data set, a collection of
mutation patterns, the leftwin, and a list of arguments to `sum`, this function
returns a matrix with columns indexed by the mutpats, the first row of which gives counts of how many of the counted mutations could be produced by each of mutpats
  and the second of which gives the number of "from" matches of the mutpats (called the "total possible").
}
\details{
in other words, for each mutation pattern a -> b
 sum the values of counts[u,v] over choices of u,v such that:
  (i) 'a' matches 'u' at some position
  (ii) in addition, 'b' matches 'v' at the same position;
store this in the first row

note that if we estimate rates by
   r.est <- countmuts(...)[1,]/countmuts(...)[2,]
then something like
   sum( r.est ) / 4
should be close to the mean density of nucleotide changes
   divergence(...)
}
