% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_tools.R
\name{countpats}
\alias{countpats}
\title{Count number of times each pattern appears in character string simseq, cyclic-ized.}
\usage{
countpats(patterns, simseq)
}
\description{
Count number of times each pattern appears in character string simseq, cyclic-ized.
}
