% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_tools.R
\name{counttrans.list}
\alias{counttrans.list}
\title{Count transitions with Shift}
\usage{
counttrans.list(lpatterns, seqlist, simseqs, leftwin, shift = 0,
  cyclic = FALSE, bases = sort(unique(unlist(strsplit(lpatterns[[1]],
  "")))))
}
\description{
count number of times each of seqlist matches corresponding lpatterns
  optionally, cyclical
if shift is nonzero, return a list  with the counts in each of (shift) frames
}
