% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_optparse.R
\name{debug.optparse}
\alias{debug.optparse}
\title{Interactive parsing of optparse-driven command line}
\usage{
debug.optparse(command, option_list)
}
\description{
Use for interactive debugging of optparse scripts
In place of e.g.  in the script this-script.R
   opt <- parse_args(OptionParser(option_list=option_list,description=usage))
use
   opt <- debug.optparse("Rscript this-script.R -a 1 -b 2 --blah", option_list)
}
