% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_residuals.R
\name{listresids}
\alias{listresids}
\title{List residuals in order}
\usage{
listresids(counts, expected, file, trim = 20,
  leftwin = (nchar(rownames(counts)[1]) - nchar(colnames(counts)[1]))/2)
}
\description{
Make a readable output ordered by z-score
 optionally writing results out to 'file'
 and trimming to only patterns with z-score above 'trim'
}
