% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{openread}
\alias{openread}
\title{Open for reading or writing}
\usage{
openread(arg)
}
\description{
Use to open stdin/stdout or process substitution things correctly
  from  http://stackoverflow.com/questions/15784373/process-substitution
}
