% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_residuals.R
\name{print.motif}
\alias{print.motif}
\title{Print motifs}
\usage{
\method{print}{motif}(pats, weights = 1, n = 24, print = TRUE,
  long = FALSE)
}
\description{
If not 'long', then positions with a base with total weight above 75% of the total are upper-case,
between 50% and 75% are lower-case, otherwise are '.'.
}
