% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_tools.R
\name{reverse.complement}
\alias{reverse.complement}
\title{Reverse Complement}
\usage{
reverse.complement(mutpats)
}
\description{
return index for each mutpat of the reverse-complement mutation pattern (or NA if none)
throw error if there are multiple matches
}
