% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_residuals.R
\name{reverse.complement.resids}
\alias{reverse.complement.resids}
\title{Collapse reverse complements in residual table}
\usage{
reverse.complement.resids(resids)
}
\arguments{
\item{resids}{A residual table, with inpat, outpat, observed, expected, resid, and z columns.}
}
\description{
Collapse reverse complements in residual table
}
