% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{selfname}
\alias{selfname}
\title{Name a list or vector with itself.}
\usage{
selfname(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
Itself, named by itself.
}
\description{
Name a list or vector with itself.
}
