% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_estimates.R
\name{divergence}
\alias{divergence}
\title{Mean Density of Nucleotide Differences}
\usage{
divergence(counts, leftwin)
}
\arguments{
\item{counts}{A tuplecounts object.}
}
\value{
The mean density.
}
\description{
Uses stringdist::stringdistmatrix.
}
